add AutoMap for news

+++ WebNews/Services/Interfaces/IGenericService.cs
@@ -4,7 +4,7 @@ public interface IGenericService<T>
 {
     IQueryable<T> GetAll();
     Task<IEnumerable<T>> GetAllAsync();
-    Task<IEnumerable<T>> GetLatestAsync(byte count);
+    Task<IEnumerable<T>> GetLatestAsync(int count);
     Task<T?> GetByIdAsync(Guid id);
     Task CreateAsync(T entity);
     Task UpdateAsync(T entity);
